Refacto code, added -a code as an API call
